backend/requirements.txt
django~=3.0.0

#common용

#개발용
django-debug-toolbar
pillow
djangorestframework